Remove old credentialAtomicQueryV2 testvector generator. Move code used by V3 to proper file.
